--- .changeset/sour-teachers-joke.md
@@ -1,8 +1,0 @@
----
-"@ledgerhq/coin-hedera": patch
-"ledger-live-desktop": patch
-"live-mobile": patch
----
-
-fix "view in explorer" button with unconfirmed hedera transactions
-support custom getTransactionExplorer in ledger-live-mobile
