+++ apps/ledger-live-desktop/src/renderer/families/hedera/index.ts
@@ -1,8 +1,8 @@
+import { getTransactionExplorer } from "@ledgerhq/live-common/families/hedera/logic";
 import AccountSubHeader from "./AccountSubHeader";
 import NoAssociatedAccounts from "./NoAssociatedAccounts";
 import sendRecipientFields from "./SendRecipientFields";
 import StepReceiveFunds from "./StepReceiveFunds";
-import getTransactionExplorer from "./getTransactionExplorer";
 import { HederaFamily } from "./types";
 
 const family: HederaFamily = {
