+++ apps/ledger-live-mobile/src/families/hedera/operationDetails.ts
@@ -1,15 +1,3 @@
-import { ExplorerView } from "@ledgerhq/types-cryptoassets";
-import { Operation } from "@ledgerhq/types-live";
-
-import { HederaOperationExtra } from "@ledgerhq/live-common/families/hedera/types";
-
-const getTransactionExplorer = (
-  explorerView: ExplorerView | null | undefined,
-  operation: Operation,
-): string | undefined => {
-  const extra = operation.extra as HederaOperationExtra;
-
-  return explorerView?.tx?.replace("$hash", extra.consensusTimestamp ?? extra.transactionId ?? "0");
-};
+import { getTransactionExplorer } from "@ledgerhq/live-common/families/hedera/logic";
 
 export default { getTransactionExplorer };
