+++ libs/coin-modules/coin-hedera/src/logic.ts
@@ -1,7 +1,7 @@
 import { ExplorerView } from "@ledgerhq/types-cryptoassets";
 import { Operation } from "@ledgerhq/types-live";
 
-import { HederaOperationExtra } from "@ledgerhq/live-common/families/hedera/types";
+import { HederaOperationExtra } from "./types";
 
 const getTransactionExplorer = (
   explorerView: ExplorerView | null | undefined,
@@ -12,4 +12,4 @@ const getTransactionExplorer = (
   return explorerView?.tx?.replace("$hash", extra.consensusTimestamp ?? extra.transactionId ?? "0");
 };
 
-export default getTransactionExplorer;
+export { getTransactionExplorer };
