+++ libs/ledger-live-common/src/families/hedera/logic.ts
@@ -1,0 +1,2 @@
+// Encapsulate for LLD & LLM
+export * from "@ledgerhq/coin-hedera/logic";
